--- src/redux/actions/index.js
@@ -1,151 +1,0 @@
-import axios from 'axios';
-
-
-//------------------------ LLAMADOS AL BACK ------------------------
-
-export function getMovies() {
-  return async (dispatch) => {
-    try {
-      const { data } = await axios.get("https://pf-henry-back.herokuapp.com/api/popular");
-
-      return dispatch({ type: "GET_MOVIES", payload: data });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-}
-
-export function getMovieDetail(id) {
-  return async (dispatch) => {
-    try {
-      const { data } = await axios.get(`https://pf-henry-back.herokuapp.com/api/${id}`);
-
-      return dispatch({ type: "GET_MOVIE_DETAIL", payload: data });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-}
-
-export function resetMovieDetail(){
-
-  return {type: 'RESET_MOVIE_DETAIL'}
-}
-
-export function getGenre(){
- return async (dispatch) => {
-  try{  
-    const { data } = await axios('https://pf-henry-back.herokuapp.com/api/genres')
-    return dispatch({ type: 'GET_GENRES', payload: data })
-  }catch (e) {
-    console.log(e)
-  }
- }
-}
-
-export function getMovieReviews(id){
-
-  return async(dispatch)=>{
-
-    try{
-
-      const { data } = await axios.get(`https://pf-henry-back.herokuapp.com/review/${id}`)
-
-      return dispatch({ type: 'GET_MOVIE_REVIEWS', payload: data })
-    }
-    catch(e){
-      console.log(e)
-    }
-
-  }
-}
-
-export function postReview(payload){
-  return async(dispatch)=>{
-
-    try{
-      await axios.post('https://pf-henry-back.herokuapp.com/review/post', payload)
-      return dispatch({type: 'POST_REVIEW'})
-    }
-    catch(e){
-      console.log(e)
-    }
-  }
-}
-
-export function getUpcoming(){
-
-  return async(dispatch)=>{
-
-    try{
-      const {data} = await axios.get(`https://pf-henry-back.herokuapp.com/api/upcoming`)
-
-      return dispatch({type: 'GET_UPCOMING', payload: data})
-    }
-    catch(e){
-      console.log(e)
-    }
-  }
-}
-
-
-//------------------------ FUNCTION ACTIONS ------------------------
-
-export function getAllFunctions() {
-  return async (dispatch) => {
-    try {
-      const { data } = await axios.get("https://pf-henry-back.herokuapp.com/function/getAll");
-
-      return dispatch({ type: "GET_FUNCTIONS", payload: data });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-}
-
-export function postFunction(payload){
-    return async(dispatch)=>{
-
-        try{
-            await axios.post('https://pf-henry-back.herokuapp.com/function/post', payload)
-            return dispatch({type:'POST_FUNCTION'})
-        }
-        catch(e){
-            console.log(e)
-        }
-
-    }
-}
-
-//------------------------ SEARCHBAR ------------------------
-
-export function getMoviesByName(payload) {
-  return {
-    type: 'GET_MOVIE_BY_NAME',
-    payload,
-  }
-}
-
-
-//------------------------ FILTERS ------------------------
-
-export function sortByRating(payload){
-  return {
-      type: 'FILTER_BY_RATING',
-      payload
-  }
-}
-
-export function sortByName(payload) {
-  return {
-    type: 'FILTER_BY_NAME',
-    payload,
-  };
-}
-
-export function sortByGenre(payload) {
-  return {
-    type: 'FILTER_BY_GENRE',
-    payload,
-  }
-}
