+++ src/redux/actions/showtimes.js
@@ -1,0 +1,36 @@
+import axios from 'axios';
+
+export function getAllShowtimes() {
+
+    return async (dispatch) => {
+
+        try {
+
+            const { data } = await axios.get("https://pf-henry-back.herokuapp.com/function/getAll");
+    
+            return dispatch({ type: "GET_FUNCTIONS", payload: data });
+        } 
+        
+        catch (e) {
+            console.log(e);
+        }
+    };
+}
+  
+export function postShowtime(payload){
+
+    return async(dispatch)=>{
+
+        try{
+
+            await axios.post('https://pf-henry-back.herokuapp.com/function/post', payload)
+            
+            return dispatch({type:'POST_FUNCTION'})
+        }
+        catch(e){
+            console.log(e)
+        }
+
+    }
+
+}
